[bug fix] replace incorrect harmonic finding code
The previous vectorized code found all peaks that were within a given
distance from a harmonic on either side. That is not what we want.
Rather, we want to find the closest peak to the theoretical harmonic
location for each harmonic (this is what findHarmonics does).

If the previous code found two peaks near the second harmonic, that
second peak would accidentally be considered the "third harmonic" peak,
which was incorrect.